home page example suggestions, and CSS suggestions
changes to the home page example to highlight the felixibilty of
the software.  also tried to make clear what is being setup in
the example.

created a class of "comment" and converted <font> to <span>

created a class of "div.code" to make code sections pretty

cannot apply the desired styling to <code> because it is not a block
level element